Add "mouse wiggle" keepalive feature.
This is a adaptation of the patch by gsomlo's github pull request
23 to upstream, which was an adaptation of my original patch from
sourceforge.  I've removed some of the code cleanup from the
previous patch, to make the code easier to keep merged with
upstream.

The original pull request was rejected as out of scope for rdesktop
but the workaround involves several external programs, such as a
Xephyr embedded X11 display and xdotool to send mouse events, and
is very cumbersome to maintain.